third_party/blink/web_tests/platform/generic/accessibility/table-headers-expected.txt

No	Month	Expenses
1	Jan	100
2	Feb	150
3	Mar	200
scope test
col head	row head	col head
col head	data	row head
row head	data	col head
row header and column header (1)
data	row head	row head
row head	row head	data
row head	data	row head
row header and column header (2)
row head	data	data
column head	column head	column head
data	data	data
scope and rowspan
row head	col head
row head
scope and colspan
col head	col head	col head
data	data	data
col head	col head
data	data	row head
col head
col head
col head
row head
data
row head
row head
row head
col head

Check whether column and row headers are reported correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS table.rowHeaderAtIndex(0).isEqual(table.cellForColumnAndRow(0, 1)) is true
PASS table.rowHeaderAtIndex(1).isEqual(table.cellForColumnAndRow(0, 2)) is true
PASS table.rowHeaderAtIndex(2).isEqual(table.cellForColumnAndRow(0, 3)) is true
PASS table.rowHeadersCount is 3
PASS table.columnHeaderAtIndex(0).isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS table.columnHeaderAtIndex(1).isEqual(table.cellForColumnAndRow(1, 0)) is true
PASS table.columnHeaderAtIndex(2).isEqual(table.cellForColumnAndRow(2, 0)) is true
PASS table.columnHeadersCount is 3
PASS table2.rowHeaderAtIndex(0).isEqual(table2.cellForColumnAndRow(1, 0)) is true
PASS table2.rowHeaderAtIndex(1).isEqual(table2.cellForColumnAndRow(2, 1)) is true
PASS table2.rowHeaderAtIndex(2).isEqual(table2.cellForColumnAndRow(0, 2)) is true
PASS table2.rowHeadersCount is 3
PASS table2.columnHeaderAtIndex(0).isEqual(table2.cellForColumnAndRow(0, 0)) is true
PASS table2.columnHeaderAtIndex(1).isEqual(table2.cellForColumnAndRow(2, 0)) is true
PASS table2.columnHeaderAtIndex(2).isEqual(table2.cellForColumnAndRow(0, 1)) is true
PASS table2.columnHeaderAtIndex(3).isEqual(table2.cellForColumnAndRow(2, 2)) is true
PASS table2.columnHeadersCount is 4
PASS table3.rowHeadersCount is 6
PASS table3.columnHeadersCount is 0
PASS table4.rowHeaderAtIndex(0).isEqual(table4.cellForColumnAndRow(0, 0)) is true
PASS table4.rowHeadersCount is 1
PASS table4.columnHeaderAtIndex(0).isEqual(table4.cellForColumnAndRow(0, 1)) is true
PASS table4.columnHeaderAtIndex(1).isEqual(table4.cellForColumnAndRow(1, 1)) is true
PASS table4.columnHeaderAtIndex(2).isEqual(table4.cellForColumnAndRow(2, 1)) is true
PASS table4.columnHeadersCount is 3
PASS table5.rowHeaderAtIndex(0).isEqual(table5.cellForColumnAndRow(0, 0)) is true
PASS table5.rowHeaderAtIndex(0).isEqual(table5.cellForColumnAndRow(0, 1)) is true
PASS table5.rowHeaderAtIndex(1).isEqual(table5.cellForColumnAndRow(1, 1)) is true
PASS table5.rowHeadersCount is 2
PASS table5.columnHeaderAtIndex(0).isEqual(table5.cellForColumnAndRow(1, 0)) is true
PASS table5.columnHeadersCount is 1
PASS table6.rowHeaderAtIndex(0).isEqual(table6.cellForColumnAndRow(2, 3)) is true
PASS table6.rowHeadersCount is 1
PASS table6.columnHeaderAtIndex(0).isEqual(table6.cellForColumnAndRow(0, 0)) is true
PASS table6.columnHeaderAtIndex(1).isEqual(table6.cellForColumnAndRow(1, 0)) is true
PASS table6.columnHeaderAtIndex(2).isEqual(table6.cellForColumnAndRow(2, 0)) is true
PASS table6.columnHeaderAtIndex(3).isEqual(table6.cellForColumnAndRow(0, 2)) is true
PASS table6.columnHeaderAtIndex(3).isEqual(table6.cellForColumnAndRow(1, 2)) is true
PASS table6.columnHeaderAtIndex(4).isEqual(table6.cellForColumnAndRow(2, 2)) is true
PASS table6.columnHeadersCount is 5
PASS table7.rowHeaderAtIndex(0).isEqual(table7.cellForColumnAndRow(0, 1)) is true
PASS table7.rowHeaderAtIndex(1).isEqual(table7.cellForColumnAndRow(2, 1)) is true
PASS table7.rowHeaderAtIndex(2).isEqual(table7.cellForColumnAndRow(0, 2)) is true
PASS table7.rowHeaderAtIndex(3).isEqual(table7.cellForColumnAndRow(1, 2)) is true
PASS table7.rowHeadersCount is 4
PASS table7.columnHeaderAtIndex(0).isEqual(table7.cellForColumnAndRow(0, 0)) is true
PASS table7.columnHeaderAtIndex(1).isEqual(table7.cellForColumnAndRow(1, 0)) is true
PASS table7.columnHeaderAtIndex(2).isEqual(table7.cellForColumnAndRow(2, 0)) is true
PASS table7.columnHeaderAtIndex(3).isEqual(table7.cellForColumnAndRow(2, 2)) is true
PASS table7.columnHeadersCount is 4
PASS successfullyParsed is true

TEST COMPLETE
